+++ config/checkstyle/checkstyle.xml
@@ -67,7 +67,6 @@
         <property name="maximum" value="0"/>
         <property name="message" value="Line has trailing spaces."/>
     </module>
-    <module name="SuppressWarningsFilter" />
 
     <!-- Checks for Headers                                -->
     <!-- See http://checkstyle.sf.net/config_header.html   -->
@@ -173,7 +172,6 @@
         <module name="TodoComment"/>
         <module name="UpperEll"/>
 
-        <module name="SuppressWarningsHolder" />
     </module>
 
 </module>
