+++ src/test/java/dk/cngroup/lentils/controller/client/AbstractClientControllerTest.java
@@ -1,4 +1,4 @@
-package dk.cngroup.lentils.controller;
+package dk.cngroup.lentils.controller.client;
 
 import dk.cngroup.lentils.entity.Cypher;
 import dk.cngroup.lentils.entity.CypherStatus;
@@ -23,7 +23,11 @@ abstract public class AbstractClientControllerTest {
     protected static final String FALSE_CODEWORD = "firefly";
 
     @Autowired
-    protected ClientController testedController;
+    protected GameController gameController;
+    @Autowired
+    protected CypherController cypherController;
+    @Autowired
+    protected HintController hintController;
     @Autowired
     protected CypherRepository cypherRepository;
     @Autowired
